fix: fix deploy issue
