+++ app/__init__.py
@@ -1,1 +1,12 @@
-from flaskscikit import app
+from flask import Flask
+from flask import request, abort
+import time
+import json
+app = Flask(__name__)
+
+@app.route("/")
+def hello():
+    return "Hello World!"
+
+def get_app():
+    return app
