--- app/flaskscikit.py
@@ -1,12 +1,0 @@
-from flask import Flask
-from flask import request, abort
-import time
-import json
-app = Flask(__name__)
-
-@app.route("/")
-def hello():
-    return "Hello World!"
-
-def get_app():
-    return app
